% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AE_relationships}
\alias{AE_relationships}
\title{The dataset of relations between the adverse events}
\format{
An object of class \code{data.frame} with 780 rows and 5 columns.
}
\usage{
AE_relationships
}
\description{
The dataset of relations between the adverse events
}
\keyword{datasets}
