% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.to.positions.R
\name{graph.to.positions}
\alias{graph.to.positions}
\title{Graph to Positions}
\usage{
graph.to.positions(allG, outputWidth, outputHeight)
}
\arguments{
\item{allG}{the graph, generated by table.to.graph}

\item{outputWidth}{the output width}

\item{outputHeight}{the output height}
}
\value{
the positions selected in the GUI
}
\description{
Generate a GUI to allow setting the positions for a graph
}
\examples{
graph.to.positions(allG, outputWidth = 1920, outputHeight = 1080)

}
